Add Http and AspNetCore projects
Note that ProblemDetails and ProblemDetailsJsonConverter were copied from the main dotnet/aspnetcore repository (and third-party notices updated). This was done because the classes are only available through a framework reference to aspnetcore - not a nuget package. Because of this, non-aspnetcore apps would basically become aspnetcore apps, which is very undesirable.