Add ErrorCodes.CaughtException
Changes the default value of the errorCode parameter in result Fail
methods - from null to a defined error code. For Fail methods with a
first parameter of type Exception, ErrorCodes.CaughtException is the new
default value for the errorCode parameter. For Fail methods with a first
parameter of type string, the default value is
ErrorCodes.InternalServerError.